I wrote tests and fixed the block problem.
Block problem:

The config parser was not checking to make sure that all blocks were terminated and that there were no '}' characters terminating blocks which had not been started.
I did this by adding a variable called block_depth, which is incremented each time a block start token is parsed and decremented each time a block end token is parsed.
An error is returned if block_depth is ever less than 0 or if block_depth is greater than 0 at the end of parsing.